requirements.txt
Flask
mysql-connector-python
Flask-SQLAlchemy
python-dotenv
transformers